requirements.txt
streamlit
firebase-admin
twilio
google-generativeai
